backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.2.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
httpx==0.27.0
structlog==24.2.0
python-dotenv==1.0.1
joblib==1.4.2
scikit-learn==1.5.1
numpy==1.26.4
lightgbm>=4.0.0
pytest>=8.2,<9
pytest-asyncio>=0.23,<0.24
